+++ CHANGELOG.md
@@ -2,6 +2,7 @@
     * refactor Github flows
     * healthcheck default scrape interval increased from 3s to 5s to prevent queue overfill
     * http status code for http failed healthchecks prepends the response
+    * fix the truncation algorithm
 
 * 0.1.10
     * upgrade all dependencies (except sysinfo)